Move Router to useRouter
